requirements.txt
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
click==8.1.7
constantly==23.10.4
cryptography==42.0.0
daphne==4.0.0
Django==5.0.1
factory-boy==3.3.0
Faker==22.6.0
graphql-core==3.2.3
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
libcst==1.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0.1
rich==13.7.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.1
strawberry-graphql==0.219.1
strawberry-graphql-django==0.30.0
Twisted==23.10.0
txaio==23.1.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
uvicorn==0.27.0.post1
zope.interface==6.1
